Removed useless test from test_binary.
The dump test was mostly there as a convenient way to be sure that
RomMap.dump didn't break. It doesn't actually unit test anything and
had to be checked by hand. Now that romtool.py is in this test
isn't doing me any good and requires an extra file that shouldn't
be in the repo anyway.