Fix map root not being found.
This was never terribly well defined. hashdb.txt is now located under
the map root, which is linked inside the romtool source. I still don't
like distributing technically-optional data files with the tool instead
of as a separate package, but python packaging makes doing it the
intuitive way damn hard.