Output no longer quotes fields
This wasn't necessary to begin with. I was trying to force excel/libre
to treat certain things as strings (most notably bitfields) rather than
integers, but quotes aren't for that. Standard csv has no indication of
types beyond content.

The only way to force a field to be treated as a string rather than a
number is to have non-numeric characters in it. The last patch did that
for bitfields and it's already the case for hex. So I don't need the
extra garbage anymore.

Removing the quotes makes command line output much cleaner as a bonus.