Add array primitive
Keeps from having to add tons of columns for things that are hardly ever
edited. I can't help but think there's a better way to do this, though.

A thought: Treat all primitives as arrays. Let `type` be a bitstring
initializer. Eliminate separate size, or use it only for arrays/structs
where type doesn't have a size in it.

This would be a pretty big change, though. Again.