Add rom identification
This relies on the datomatic no-intro rom databases. There were several
possible approaches, none of them great:

1. Download and cache them on first run. This would be ideal, but I
   couldn't find a good way to do it. Getting to them requires jumping
   through some anti-bot hoops.
2. Have the user download them to somewhere romtool can look for them.
   No, no one will actually do that.
3. Include them verbatim in the source repo. I'd prefer not to do that,
   plus I have no idea if licensing would be an issue.
4. Generate a tsv from the dat with just the data I need for
   identification, and put the tsv in the repo. The file is still huge,
   larger than everything else put together.
5. Generate such a tsv at build time, but don't commit it. Find a way to
   make tests fail if it's missing, so a release can't accidentally remove
   it (maybe check its hash vs. a known good hash that *is* committed?)

I'm going with #5 as the least-bad option. For now.

The included xsd schema for the dat files isn't actually used, yet, but
might be in future. I don't want to lose it given how aggravating it was
to make it work.