Check python optimization status
There are a couple of expensive asserts in anytree and BitArrayView that
slow things down.  They're mostly noticeable when dumping/loading a
directory.  Optimization skips asserts, so check if it's on and say
something if needed.

I'm not sure whether to treat this as a warning or not. Debug mode is
non-optimal, but it's also the default. Treating as info for now.