Fix crash setting entity attributes
For future reference, can be reproed on the previous commit by building
from a changeset. _setattr was defined such that the partialed attribute
name's position conflicted with the value being set by the caller.
(essentially I tried to match __setattr__'s call signature, but that was
neither necessary nor appropriate for this usage)