Move the structure type registry into the Rom
The current class attribute is a quasi-global, which makes lots of
things difficult. It's impossible to open more than one Rom map per
run right now because they use the same namespace. It also requires
explicit teardown in tests.

Just get rid of it and make it a rom instance attribute.