Allow tables not associated with an entity
Note that with the table-id fallback removed, entity ID must now always
be explicit. This is slightly more annoying, since in many cases it will
be the same. It also adds an extra step to crossref resolution. But it
allows maps to define tables that aren't dumped, e.g.  because they
overlap with other tables.

The motivating case was a loot table that collectively covered several
other tables of differently-typed objects, which could not be dumped
together.