Move string manipulation into the field class
The string methods on a view were warts -- they needed to know field
attributes unnecessary for any other type.

This simplifies table lookups too, but that could still use more work.
The current field interface feels not-quite-right; it has no good way of
dealing with cases where the reader is not a structure.