Improve reporting of map errors
Mainly, this fixes cases where KeyErrors resulting from map construction
propagate further than they should. The motivating meta-error was a bad
value for a field's 'unit' column generating a KeyError (from
Unit.__getitem__) that got caught by an unrelated ChainMap used for map
detection. The resulting error message was extremely confusing;
ChainMap checks each underlying mapping, and raises its own KeyError if
all of them fail -- but it doesn't distinguish between a submap that
doesn't have the key and unrelated KeyErrors further down the stack.
In the map detection case, it ended with detection reporting that the
rom hash was not found -- when the actual error was a bogus Unit key in
the map it did find.

Probably the design should be revisited, but I'm not sure exactly what
change to make.