Use explicit `i` in CalculatedIndex.EvalContext
The next use-case I've found for it (pokedex names in Ruby) breaks on
the current implicit i (it needs to map i to i+1). Just put the index
and table objects directly in context. The definitions will be more
verbose, but working and ugly is better than broken and clean.