Fix bad error handling during table and eset iteration
In some cases, an IndexError raised by a bad view slice (e.g. because of
a bogus entry in an in-rom index) propagates out of Table.__getitem__.
This looks to callers like the index they asked for is out-of-bounds for
the table itself, which causes weirdness such as Sequence.__iter__
treating a bad index as end-of-sequence.

That shouldn't happen. I'm not sure what *should* happen, but that
behavior is definitely wrong. Replace such exceptions with ValueError
for now, until I come up with something better.