Remove bogus(?) check
I have no idea WTF this was supposed to do. As far as I can tell
it originally looked up the string in the crossrefed table, then looked up
that value in rom.entities/rom.tables (whichever subset contained the
crossref target), and skipped the write if the original string matched
that table's name.

So `if entities[locate(xref, string)].name == string`, do a no-op.

...I have no idea what that could possibly trigger such a case.